Inline and remove POutputValueBase
