+++ sdks/java/core/src/main/java/org/apache/beam/sdk/values/PDone.java
@@ -21,13 +21,20 @@ import java.util.Collections;
 import java.util.Map;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.io.WriteFiles;
+import org.apache.beam.sdk.transforms.AppliedPTransform;
 import org.apache.beam.sdk.transforms.PTransform;
 
 /**
  * {@link PDone} is the output of a {@link PTransform} that has a trivial result,
  * such as a {@link WriteFiles}.
  */
-public class PDone extends POutputValueBase {
+public class PDone implements POutput {
+
+  private final Pipeline pipeline;
+
+  private PDone(Pipeline pipeline) {
+    this.pipeline = pipeline;
+  }
 
   /**
    * Creates a {@link PDone} in the given {@link Pipeline}.
@@ -37,12 +44,25 @@ public class PDone extends POutputValueBase {
   }
 
   @Override
+  public Pipeline getPipeline() {
+    return pipeline;
+  }
+
+  /**
+   * A {@link PDone} contains no {@link PValue PValues}.
+   */
+  @Override
   public Map<TupleTag<?>, PValue> expand() {
-    // A PDone contains no PValues.
     return Collections.emptyMap();
   }
 
-  private PDone(Pipeline pipeline) {
-    super(pipeline);
-  }
+  /** Does nothing; there are no concrete outputs to record. */
+  @Override
+  public void recordAsOutput(AppliedPTransform<?, ?, ?> transform) {}
+
+  /**
+   * Does nothing; there is nothing to finish specifying.
+   */
+  @Override
+  public void finishSpecifyingOutput(PInput input, PTransform<?, ?> transform) { }
 }
